fix: doxygen; rm not needed function; consolidate file
