+++ Makefile
@@ -6,7 +6,7 @@
 #    By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/07/28 13:03:05 by gwolf             #+#    #+#              #
-#    Updated: 2023/08/11 17:16:05 by gwolf            ###   ########.fr        #
+#    Updated: 2023/08/11 19:28:20 by gwolf            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -88,7 +88,6 @@ SRC :=	utils_memory.c \
 		syntax.c \
 		expand_expander.c \
 		expand_field_split.c \
-		expand_quote_removal.c \
 		expand_tilde.c \
 		expand_utils.c \
 		expand_dollar.c \
