+++ inc/mod_expand.h
@@ -6,7 +6,7 @@
 /*   By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/07 13:08:04 by gwolf             #+#    #+#             */
-/*   Updated: 2023/08/11 18:31:43 by gwolf            ###   ########.fr       */
+/*   Updated: 2023/08/11 19:27:56 by gwolf            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,6 +59,8 @@ t_err	ft_expand_arg(t_tkn_list **list, t_hashtable *symtab);
 
 //expand_expander.c
 t_err	ft_expander_arg(t_track *input, t_hashtable *symtab, t_type type);
+t_err	ft_rm_single_quote(t_track *input);
+t_err	ft_rm_double_quote(t_track *input);
 
 //expand_tilde.c
 t_err	ft_expand_tilde(t_track *input, t_hashtable *symtab);
@@ -66,7 +68,7 @@ t_err	ft_get_tilde_replace(t_str token, t_hashtable *symtab,
 			t_str *replace, size_t *pos);
 t_err	ft_get_tilde_var(t_track *input, t_str *token);
 
-//expand_var.c
+//expand_dollar.c
 t_err	ft_expand_dollar(t_track *input, t_hashtable *symtab);
 t_err	ft_special_dollar(t_str *token, t_str *replace);
 t_err	ft_get_dollar_var(t_track *input, t_str *token);
@@ -78,11 +80,6 @@ t_err	ft_count_expand_words(t_track *input, size_t *words, int last_expand_len);
 t_err	ft_split_node(t_track *input, t_tkn_list **cur_node, t_buf *buf);
 t_err	ft_tokenise_fs(t_src *src, t_tok *token, t_buf *buf, t_track *input);
 t_err	ft_partition_fs(t_src *src, t_buf *buf);
-
-//expand_quote_removal.c
-t_err	ft_quote_removal(char *str);
-t_err	ft_rm_single_quote(t_track *input);
-t_err	ft_rm_double_quote(t_track *input);
 
 //expand_utils.c
 t_err	ft_init_tracker(t_track *input, char *str);
